% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_read_excel.R
\name{multi_read_excel}
\alias{multi_read_excel}
\title{my_fun A function to read all available excel files in a folder}
\usage{
multi_read_excel(folder = "data/")
}
\arguments{
\item{A}{path to a folder (type = character)}
}
\value{
A list of data frames, after reading all available excel files in a folder
}
\description{
my_fun A function to read all available excel files in a folder
}
\examples{
multi_read_excel(folder = "inst/")
}
