% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possi_excel.R
\name{possi_excel}
\alias{possi_excel}
\title{my_fun A function to read an excel file and return NULL if there is a problem with the file}
\usage{
possi_excel(...)
}
\arguments{
\item{A}{filename or pathfile (type = character)}
}
\value{
A data frame when the file is read successfully, or NULL if there is an error
}
\description{
my_fun A function to read an excel file and return NULL if there is a problem with the file
}
