% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.exam-package.R
\docType{package}
\name{practice.exam-package}
\alias{practice.exam}
\alias{practice.exam-package}
\title{practice.exam: Build A Package From Rmarkdown File}
\description{
Package for practicing exam for package creation.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
